Generic directory function file created and successfully implemented
