+++ plotWHAN.py
@@ -7,14 +7,14 @@ import numpy as np
 import matplotlib.pyplot as plt
 from astropy.io import fits
 from matplotlib import rc
-import os
+import direcFuncs
 
 
 def plotWHAN(filename):
 
     temp = fits.open(filename)
 
-    nFP = setupNewDir(filename, "")
+    nFP = direcFuncs.setupNewDir(filename, "WHAN", "")
 
     # for splitting the filename
     name = (filename.split('/')[-1]).split('.')[0]
@@ -57,25 +57,3 @@ def plotWHAN(filename):
     plt.savefig(nFP + name_plateNum_Bundle + '_WHAN.png')
     # plt.show()
     plt.close()
-
-
-def setupNewDir(filename, typeStr):
-    # ensures new folder
-    # sets up folder path for new files
-    fileLs = filename.split('/')
-    # newFldrNme = fileLs[-1]
-    del fileLs[-1]
-
-    newFldrPath = '/'.join(fileLs) + '/Figures/WHAN' + typeStr + '/'
-    assure_path_exists(newFldrPath)
-
-    return newFldrPath
-
-
-def assure_path_exists(path):
-    # from
-    # https://justgagan.wordpress.com/2010/09/22/python-create-path-or-directories-if-not-exist/
-    dir = os.path.dirname(path)
-    if not os.path.exists(dir):
-        os.makedirs(dir)
-        print("New directory made {" + dir + "/}")
